# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Troubleshoot Stack API",
    "version": "0.1.0",
    "description": "MVP API for triage and explain workflows."
  },
  "servers": [
    {
      "url": "https://{restapi_id}.execute-api.{region}.amazonaws.com/{stage}",
      "variables": {
        "restapi_id": { "default": "example" },
        "region": { "default": "us-west-2" },
        "stage": { "default": "dev" }
      }
    }
  ],
  "paths": {
    "/triage": {
      "post": {
        "summary": "Analyze an error log or trace stack",
        "operationId": "triage",
        "security": [ { "ApiKeyAuth": [] } ],
        "x-amazon-apigateway-integration": {
          "type": "http_proxy",
          "httpMethod": "POST",
          "connectionType": "INTERNET",
          "uri": "http://${alb_dns_name}:${alb_listener_port}/triage"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TriageRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Triage response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TriageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" },
          "500": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/explain": {
      "post": {
        "summary": "Explain a hypothesis or follow-up question",
        "operationId": "explain",
        "security": [ { "ApiKeyAuth": [] } ],
        "x-amazon-apigateway-integration": {
          "type": "http_proxy",
          "httpMethod": "POST",
          "connectionType": "INTERNET",
          "uri": "http://${alb_dns_name}:${alb_listener_port}/explain"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ExplainRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Explain response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TriageResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" },
          "500": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/status": {
      "get": {
        "summary": "Service health status",
        "operationId": "status",
        "x-amazon-apigateway-integration": {
          "type": "http_proxy",
          "httpMethod": "GET",
          "connectionType": "INTERNET",
          "uri": "http://${alb_dns_name}:${alb_listener_port}/status"
        },
        "responses": {
          "200": {
            "description": "Status response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "responses": {
      "Problem": {
        "description": "RFC 7807 Problem Details",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/ProblemDetails" }
          }
        }
      }
    },
    "schemas": {
      "TriageRequest": {
        "type": "object",
        "required": ["source", "payload"],
        "properties": {
          "id": { "type": "string" },
          "source": { "type": "string", "description": "Origin of the log or trace." },
          "payload": { "type": "string", "description": "Raw error log or trace stack." },
          "timestamp": { "type": "string", "format": "date-time" },
          "context": { "type": "object", "additionalProperties": true },
          "conversation_id": { "type": "string" },
          "prior_messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PriorMessage" }
          }
        }
      },
      "ExplainRequest": {
        "type": "object",
        "required": ["payload"],
        "properties": {
          "id": { "type": "string" },
          "payload": { "type": "string", "description": "Follow-up question or hypothesis id." },
          "timestamp": { "type": "string", "format": "date-time" },
          "context": { "type": "object", "additionalProperties": true },
          "conversation_id": { "type": "string" },
          "prior_messages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PriorMessage" }
          }
        }
      },
      "PriorMessage": {
        "type": "object",
        "required": ["role", "content"],
        "properties": {
          "role": { "type": "string", "enum": ["user", "assistant", "system"] },
          "content": { "type": "string" }
        }
      },
      "TriageResponse": {
        "type": "object",
        "required": ["request_id", "timestamp", "hypotheses", "runbook_steps"],
        "properties": {
          "request_id": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "hypotheses": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Hypothesis" }
          },
          "runbook_steps": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RunbookStep" }
          },
          "proposed_fix": {
            "type": "string"
          },
          "risk_notes": { "type": "array", "items": { "type": "string" } },
          "rollback": { "type": "array", "items": { "type": "string" } },
          "next_checks": { "type": "array", "items": { "type": "string" } },
          "metadata": {
            "type": "object",
            "properties": {
              "model_version": { "type": "string" },
              "tool_versions": { "type": "object", "additionalProperties": true },
              "cache_hit": { "type": "boolean" },
              "latency_ms": { "type": "number" },
              "prompt_version": { "type": "string" },
              "prompt_filename": { "type": "string" },
              "model_id": { "type": "string" },
              "token_usage": {
                "type": "object",
                "properties": {
                  "prompt_tokens": { "type": "integer" },
                  "completion_tokens": { "type": "integer" },
                  "total_tokens": { "type": "integer" },
                  "generated_at_ms": { "type": "integer" }
                },
                "additionalProperties": true
              },
              "guardrails": {
                "type": "object",
                "properties": {
                  "citation_missing_count": { "type": "integer" },
                  "redactions": { "type": "integer" },
                  "issues": { "type": "array", "items": { "type": "string" } }
                },
                "additionalProperties": true
              },
              "category": { "type": "string" },
              "recommended_tool_calls": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ToolCall" }
              }
            },
            "additionalProperties": true
          },
          "conversation_id": { "type": "string" }
        }
      },
      "Hypothesis": {
        "type": "object",
        "required": ["id", "rank", "confidence", "explanation"],
        "properties": {
          "id": { "type": "string" },
          "rank": { "type": "integer", "minimum": 1 },
          "confidence": { "type": "number", "minimum": 0, "maximum": 1 },
          "explanation": { "type": "string" },
          "citations": { "type": "array", "items": { "$ref": "#/components/schemas/EvidenceMapEntry" } }
        }
      },
      "EvidenceMapEntry": {
        "type": "object",
        "required": ["source_type", "source_id", "line_start", "line_end", "excerpt_hash"],
        "properties": {
          "source_type": { "type": "string", "description": "log|tool" },
          "source_id": { "type": "string" },
          "line_start": { "type": "integer" },
          "line_end": { "type": "integer" },
          "excerpt_hash": { "type": "string" }
        }
      },
      "ToolCall": {
        "type": "object",
        "required": ["tool"],
        "properties": {
          "tool": { "type": "string" },
          "call_spec": { "type": "object", "additionalProperties": true }
        }
      },
      "RunbookStep": {
        "type": "object",
        "required": ["step_number", "description"],
        "properties": {
          "step_number": { "type": "integer", "minimum": 1 },
          "description": { "type": "string" },
          "command_or_console_path": { "type": "string" },
          "estimated_time_mins": { "type": "number" }
        }
      },
      "StatusResponse": {
        "type": "object",
        "required": ["status", "timestamp", "dependencies"],
        "properties": {
          "status": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "dependencies": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["name", "healthy"],
              "properties": {
                "name": { "type": "string" },
                "healthy": { "type": "boolean" },
                "detail": { "type": "string" }
              }
            }
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "required": ["type", "title", "status", "detail", "instance"],
        "properties": {
          "type": { "type": "string", "format": "uri" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" }
        }
      }
    }
  }
}